refactor(phase4b): extract ApiError and RateLimitError to config/api/errors.js
Extracted reusable error classes from the 406-line api.js config module
into self-contained src/config/api/errors.js. Reduces api.js by 20 lines
and makes error classes importable without pulling in the Axios instance.
